weather card changes,tab changes

+++ Airline/XAirlines/Helpers/CurrencyHelper.cs
@@ -52,7 +52,7 @@ namespace Airlines.XAirlines.Helpers
         public float USDBGN { get; set; }
         public float USDBHD { get; set; }
         public float USDBIF { get; set; }
-        public int USDBMD { get; set; }
+        public float USDBMD { get; set; }
         public float USDBND { get; set; }
         public float USDBOB { get; set; }
         public float USDBRL { get; set; }
@@ -61,7 +61,7 @@ namespace Airlines.XAirlines.Helpers
         public float USDBTN { get; set; }
         public float USDBWP { get; set; }
         public float USDBYN { get; set; }
-        public int USDBYR { get; set; }
+        public float USDBYR { get; set; }
         public float USDBZD { get; set; }
         public float USDCAD { get; set; }
         public float USDCDF { get; set; }
@@ -71,7 +71,7 @@ namespace Airlines.XAirlines.Helpers
         public float USDCNY { get; set; }
         public float USDCOP { get; set; }
         public float USDCRC { get; set; }
-        public int USDCUC { get; set; }
+        public float USDCUC { get; set; }
         public float USDCUP { get; set; }
         public float USDCVE { get; set; }
         public float USDCZK { get; set; }
@@ -185,7 +185,7 @@ namespace Airlines.XAirlines.Helpers
         public float USDTZS { get; set; }
         public float USDUAH { get; set; }
         public float USDUGX { get; set; }
-        public int USDUSD { get; set; }
+        public float USDUSD { get; set; }
         public float USDUYU { get; set; }
         public float USDUZS { get; set; }
         public float USDVEF { get; set; }
